Revert "Extract ExecutionPlanRunner profiling helpers to dedicated partial fi…"
This reverts commit add58963beac044707dd7ecae8124d01f7ad0b46.
